+++ pkgs/code_builder/lib/src/emitter.dart
@@ -188,7 +188,12 @@ class DartEmitter extends Object
     }
     out.write('mixin ${spec.name}');
     visitTypeParameters(spec.types.map((r) => r.type), out);
-    if (spec.on != null) {
+    if (spec.onTypes.isNotEmpty) {
+      out.write(' on ');
+      visitAll<Reference>(spec.onTypes, out, (r) {
+        r.type.accept(this, out);
+      }, ',');
+    } else if (spec.on != null) {
       out.write(' on ');
       spec.on!.type.accept(this, out);
     }