+++ pkgs/unified_analytics/CHANGELOG.md
@@ -1,7 +1,23 @@
-## 8.0.16
+## 8.0.18
 
+- Added optional `hostArch` parameter to `Event.flutterCommandResult`.
+- Added optional `pubspecHasFlutterSdk`, `pubspecEnvironmentSdk`, and
+  `pubspecDependencies` parameters to the `Event.dartCliCommandExecuted`
+  constructor.
+- Dependencies are deterministically sorted and chunked using a hash-based
+  algorithm salted with the canonical dependency set to fit within Google
+  Analytics 4 parameter limitations without alphabetical or global package bias.
+- Add optional `minSdkConstraints` and `languageVersionOverrides` parameters
+  to the `Event.contextStructure` constructor.
+- Add `isExternal` to indicate if analytics are collected from an extenal build
+  system. The analytics consent mechanism is bypassed for non external builds.
+
+## 8.0.17
+- Added optional `agentPlugin` parameter to the `Event.dartMCPEvent` constructor to identify 
+agent plugin usage.
+
+## 8.0.16
 - Added `Event.packageSkillsEvent` to track events from package:skills.
-- Run `dart format`.
 
 ## 8.0.15
 - Added IDE and plugin information to `Event.serverSession`.